+++ R/classAchillesTable.R
@@ -106,8 +106,8 @@ validateAchillesTable <- function(table,
   assertClass(table, c("achilles_table", "cdm_table"), nm = nm, call = call)
   assertTable(table, empty = empty, nm = nm, call = call)
   assertChoice(
-    version, choices = supportedCdmVersions, null = TRUE, length = 1,
-    call = call
+    version, choices = omopgenerics::supportedCdmVersions, null = TRUE,
+    length = 1, call = call
   )
   assertLogical(cast, length = 1, call = call)
   if (is.null(version)) {